Reorganize candidate graph tests
