+++ tests/test_candidate_graph/test_compute_graph.py
@@ -1,0 +1,55 @@
+from collections import Counter
+
+import pytest
+from motile_toolbox.candidate_graph import (
+    EdgeAttr,
+    graph_from_segmentation,
+)
+
+
+def test_graph_from_segmentation_2d(segmentation_2d, graph_2d):
+    # test with 2D segmentation
+    cand_graph = graph_from_segmentation(
+        segmentation=segmentation_2d,
+        max_edge_distance=100,
+        iou=True,
+    )
+    assert Counter(list(cand_graph.nodes)) == Counter(list(graph_2d.nodes))
+    assert Counter(list(cand_graph.edges)) == Counter(list(graph_2d.edges))
+    for node in cand_graph.nodes:
+        assert Counter(cand_graph.nodes[node]) == Counter(graph_2d.nodes[node])
+    for edge in cand_graph.edges:
+        print(cand_graph.edges[edge])
+        assert (
+            pytest.approx(cand_graph.edges[edge][EdgeAttr.DISTANCE.value], abs=0.01)
+            == graph_2d.edges[edge][EdgeAttr.DISTANCE.value]
+        )
+        assert (
+            pytest.approx(cand_graph.edges[edge][EdgeAttr.IOU.value], abs=0.01)
+            == graph_2d.edges[edge][EdgeAttr.IOU.value]
+        )
+
+    # lower edge distance
+    cand_graph = graph_from_segmentation(
+        segmentation=segmentation_2d,
+        max_edge_distance=15,
+    )
+    assert Counter(list(cand_graph.nodes)) == Counter(["0_1", "1_1", "1_2"])
+    assert Counter(list(cand_graph.edges)) == Counter([("0_1", "1_2")])
+    assert cand_graph.edges[("0_1", "1_2")][EdgeAttr.DISTANCE.value] == pytest.approx(
+        11.18, abs=0.01
+    )
+
+
+def test_graph_from_segmentation_3d(segmentation_3d, graph_3d):
+    # test with 3D segmentation
+    cand_graph = graph_from_segmentation(
+        segmentation=segmentation_3d,
+        max_edge_distance=100,
+    )
+    assert Counter(list(cand_graph.nodes)) == Counter(list(graph_3d.nodes))
+    assert Counter(list(cand_graph.edges)) == Counter(list(graph_3d.edges))
+    for node in cand_graph.nodes:
+        assert Counter(cand_graph.nodes[node]) == Counter(graph_3d.nodes[node])
+    for edge in cand_graph.edges:
+        assert pytest.approx(cand_graph.edges[edge], abs=0.01) == graph_3d.edges[edge]
